Move maintainence thread func take u64 rather than int
signed rollover is UB

+++ vendored/libfuse/include/maintenance_thread.hpp
@@ -1,5 +1,7 @@
 #pragma once
 
+#include "base_types.h"
+
 #include <functional>
 #include <vector>
 
@@ -9,5 +11,5 @@ class MaintenanceThread
 public:
   static void setup();
   static void stop();
-  static void push_job(const std::function<void(int)> &func);
+  static void push_job(const std::function<void(u64)> &func);
 };
